+++ lncrawl/services/sources/tester.py
@@ -2,6 +2,7 @@ import hashlib
 import types
 from typing import Any, Callable, List, Type
 
+from lncrawl.context import ctx
 from lncrawl.core import Chapter, Crawler, Novel, PageSoup
 from lncrawl.services.sources.helper import extract_crawlers
 from lncrawl.utils.log_sink import LogSink
@@ -87,7 +88,13 @@ def run_crawler_test(
     crawler_log_sink.attach(emit)
 
     step("Initializing crawler")
-    crawler = constructor(origin=origin)
+    crawler = constructor(
+        origin=origin,
+        scraper=ctx.scraper.open(
+            origin,
+            rate_limit=constructor.request_rate_limit,
+        ),
+    )
     crawler.initialize()
     meta("origin", origin)